MINOR: Tidy polyblend on classic

+++ client.h
@@ -225,7 +225,7 @@ typedef struct
 typedef struct 
 {
 	int		destcolor[3];
-	int		percent;		// 0-256
+	float	percent;		// 0-256
 } cshift_t;
 
 #define	CSHIFT_CONTENTS	0
